Refactored Scoreboard
