--- code/ui/NewScoreboard.cs
@@ -1,162 +1,0 @@
-using System.ComponentModel;
-
-using Sandbox;
-using Sandbox.UI;
-using Sandbox.UI.Construct;
-using System.Collections.Generic;
-
-namespace TTTReborn.UI
-{
-    public class Scoreboard : Panel
-    {
-        public Scoreboard()
-        {
-            StyleSheet.Load("/ui/NewScoreboard.scss");
-            header = new Header();
-            AddHeader();
-            AddTableHeader();
-            var mainContent = Add.Panel("mainContent");
-            // TODO: create playergroups & for each playergroup:
-            AddScoreboardGroup(mainContent);
-
-            PlayerScore.OnPlayerAdded += AddPlayer;
-            PlayerScore.OnPlayerRemoved += RemovePlayer;
-            Add.Panel("footer");
-            // TODO: Implement UpdatePlayer method
-            // PlayerScore.OnPlayerUpdated += UpdatePlayer;
-
-            // why is: var (item, index) ; throwing an error here?
-            foreach (var player in PlayerScore.All)
-            {
-                AddPlayer(player);
-            }
-        }
-
-        public class Header : Panel
-        {
-            public void UpdateServerInfo()
-            {
-                Log.Info("asdasd");
-                this.ServerInfo.Text = this.GetServerInfoStr();
-            }
-            public string GetServerInfoStr()
-            {
-                // TODO: Get this out of the header
-                // TODO: Fill the other variables
-                return $"{PlayerScore.All.Length}/MAXPLAYER - MAPNAME - ROUND/TIME LEFT";
-            }
-            public Panel ScoreboardLogo;
-            public Panel InformationHolder;
-            public Label ServerName;
-            public Label ServerInfo;
-            public Label ServerDescription;
-            public Panel Canvas;
-        }
-        public class ScoreboardGroup : Panel
-        {
-            // TODO: Implement logic for the player counter in the title
-            public Label GroupTitle;
-            public Panel GroupContent;
-            public Panel Canvas;
-
-
-        }
-
-        public Dictionary<int, ScoreboardEntry> Entries = new();
-
-        public Dictionary<string, ScoreboardGroup> ScoreboardGroups = new();
-        public Header header;
-
-        protected void AddHeader()
-        {
-            header.Canvas = Add.Panel("header");
-            header.ScoreboardLogo = header.Canvas.Add.Panel("scoreboardLogo");
-            header.InformationHolder = header.Canvas.Add.Panel("informationHolder");
-            header.ServerName = header.InformationHolder.Add.Label("Here will be the servername","serverName");
-            header.ServerInfo = header.InformationHolder.Add.Label(header.GetServerInfoStr(), "serverInfo");
-            header.ServerDescription = header.InformationHolder.Add.Label("This is the server description: Lorem ipsum dolor sit  elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat", "serverDescription");
-        }
-
-        protected void AddTableHeader()
-        {
-            var tableHeader = Add.Panel("tableHeader");
-            tableHeader.Add.Label("3 Innocents left", "name");
-            tableHeader.Add.Label("Karma", "karma");
-            tableHeader.Add.Label("Score", "score");
-            tableHeader.Add.Label("Ping", "ping");
-        }
-
-        protected void AddScoreboardGroup(Panel Content)
-        {
-            string group = "alive";
-            var scoreboardGroup = new ScoreboardGroup
-            {
-
-            };
-            scoreboardGroup.Canvas = Content.Add.Panel("scoreboardGroup alive");
-            scoreboardGroup.GroupTitle = scoreboardGroup.Canvas.Add.Label(group.ToUpper() + " - 0", "scoreboardGroup__title");
-            scoreboardGroup.GroupContent = scoreboardGroup.Canvas.Add.Panel("scoreboardGroup__content");
-
-            ScoreboardGroups[group] = scoreboardGroup;
-        }
-
-        protected void AddPlayer(PlayerScore.Entry entry)
-        {
-            // TODO: Get proper scoreboardGroup for entry
-            string group = "alive";
-            var section = ScoreboardGroups[group].GroupContent;
-            var p = section.AddChild<ScoreboardEntry>();
-            bool isOdd = (PlayerScore.All.Length % 2) != 0;
-            p.UpdateFrom(entry, isOdd);
-            header.UpdateServerInfo();
-            Entries[entry.Id] = p;
-        }
-        protected void RemovePlayer(PlayerScore.Entry entry)
-        {
-            if (Entries.TryGetValue(entry.Id, out var panel))
-            {
-                panel.Delete();
-                Entries.Remove(entry.Id);
-            }
-        }
-        public override void Tick()
-        {
-            base.Tick();
-
-            SetClass("open", Local.Client?.Input.Down(InputButton.Score) ?? false);
-        }
-    }
-
-    public class ScoreboardEntry : Panel
-    {
-        public PlayerScore.Entry Entry;
-
-        public Label PlayerName;
-        public Label Karma;
-        public Label Score;
-        public Label Ping;
-
-        public ScoreboardEntry()
-        {
-            AddClass("entry");
-
-            PlayerName = Add.Label("PlayerName", "name");
-            Karma = Add.Label("", "karma");
-            Score = Add.Label("", "score");
-            Ping = Add.Label("", "ping");
-        }
-
-        public virtual void UpdateFrom(PlayerScore.Entry entry, bool isOdd)
-        {
-            Entry = entry;
-
-            PlayerName.Text = entry.GetString("name");
-            Karma.Text = entry.Get<int>("karma", 0).ToString();
-            Score.Text = entry.Get<int>("score", 0).ToString();
-            Ping.Text = entry.Get<int>("ping", 0).ToString();
-
-            SetClass("me", Local.Client != null && entry.Get<ulong>("steamid", 0) == Local.Client.SteamId);
-            SetClass(isOdd ? "odd" : "even", true);
-        }
-    }
-}
