--- code/ui/NewScoreboard.scss
@@ -1,130 +1,0 @@
-Scoreboard {
-    position: absolute;
-    left: 50%;
-    top: 50%;
-    font-family: Poppins;
-    border-radius: 8px;
-    width: 1200px;
-    transform: translateX(-50%) translateY(-60%);
-    border: 1px solid #000;
-    color: #fafafa;
-    opacity: 0;
-    z-index: 100;
-    transition: opacity 0.05s linear;
-    display: flex;
-    flex-direction: column;
-    background-color: #0009;
-
-    &.open {
-        opacity: 1;
-    }
-
-    .header {
-        padding: 16px 30px;
-        padding-bottom: 25px;
-        border-bottom: #363636 1px solid;
-        background-color: #0007;
-        border-top-left-radius: 7px;
-        border-top-right-radius: 7px;
-        display: flex;
-
-        .scoreboardLogo {
-            background-image: url( /assets/TTT-Reborn-Logo.png );
-            background-size: 100%;
-            width: 100px;
-            height: 100px;
-            margin-right: 20px;
-            // flex-basis:100px;
-        }
-        // margin-bottom: 25px;
-
-        .informationHolder {
-            display: flex;
-            flex-direction: column;
-            .serverName {
-                font-size: 24px;
-                font-weight: 500;
-            }
-            .serverInfo {
-                font-weight: 400;
-                font-size: 18px;
-            }
-            .serverDescription {
-                font-weight: 300;
-                font-size: 0.9em;
-                margin-top: 10px;
-                margin-bottom: 0;
-                padding-right: 75px;
-            }
-        }
-    }
-    .tableHeader {
-        background-color: #0007;
-        display: flex;
-        flex-direction: row;
-        font-weight: 500;
-        font-size: 20px;
-        padding-bottom: 8px;
-        .name {
-            padding-left: 20px;
-        }
-    }
-    .mainContent {
-        .scoreboardGroup {
-            display: flex;
-            flex-direction: column;
-            width: 100%;
-            margin-top: 5px;
-
-            .scoreboardGroup__title {
-                background-color: #3f7e51;
-                font-size: 14px;
-                font-weight: 300;
-                width: 100px;
-                border-top-right-radius: 4px;
-                padding: 0 10px 0 20px;
-            }
-            .scoreboardGroup__content {
-                // border-image: linear-gradient(to bottom, #3f7e51, #3f7e51, #3f7e51, rgba(0, 0, 0, 0)) 1 1%;
-                border: 2px solid #3f7e51;
-                // border-color: #3f7e51;
-                border-right-width: 0;
-                border-left-width: 1px;
-                border-bottom-width: 0;
-                display: flex;
-                flex-direction: column;
-                box-shadow: 0px 3px 5px -1px #00000040;
-
-                ScoreboardEntry {
-                    width: 100%;
-                    padding-left: 20px;
-                }
-                .odd {
-                    background-color: #0a0a0a66;
-                }
-                .even {
-                    background-color: #2a2a2a66;
-                }
-            }
-        }
-    }
-
-    .name {
-        flex-grow: 1;
-    }
-    .score,
-    .karma,
-    .ping {
-        width: 90px;
-        text-align: center;
-    }
-
-    .footer {
-        margin-top:10px;
-        border-bottom-left-radius: 7px;
-        border-bottom-right-radius: 7px;
-        border-top: #363636 1px solid;
-        background-color: #0007;
-        padding: 16px 30px;
-    }
-}
